refactor: move selectors into store directory
