+++ dev-client/src/screens/LocationScreens/LocationSoilIdScreen.tsx
@@ -18,7 +18,7 @@
 import {useTranslation} from 'react-i18next';
 import {ScrollView} from 'react-native-gesture-handler';
 
-import {selectSite} from 'terraso-mobile-client/model/selectors';
+import {selectSite} from 'terraso-mobile-client/store/selectors';
 import {Coords} from 'terraso-client-shared/types';
 
 import {Box} from 'terraso-mobile-client/components/NativeBaseAdapters';
