+++ dev-client/src/screens/ProjectInputScreen/SoilPitSettings.tsx
@@ -20,7 +20,7 @@ import {useTranslation} from 'react-i18next';
 
 import {Button} from 'native-base';
 
-import {useProjectSoilSettings} from 'terraso-mobile-client/model/selectors';
+import {useProjectSoilSettings} from 'terraso-mobile-client/store/selectors';
 import {
   DEPTH_PRESETS,
   LabelledDepthInterval,
