+++ dev-client/src/screens/SlopeScreen/SlopeShapeScreen.tsx
@@ -28,7 +28,7 @@ import {
 import {
   selectSoilData,
   selectUserRoleSite,
-} from 'terraso-mobile-client/model/selectors';
+} from 'terraso-mobile-client/store/selectors';
 import {updateSoilData} from 'terraso-mobile-client/model/soilId/soilIdSlice';
 
 import ConcaveConcave from 'terraso-mobile-client/assets/slope/shape/concave-concave.svg';
