+++ dev-client/src/screens/SlopeScreen/components/ManualSteepnessModal.tsx
@@ -22,7 +22,7 @@ import {Formik} from 'formik';
 import {Button} from 'native-base';
 import * as yup from 'yup';
 
-import {selectSoilData} from 'terraso-mobile-client/model/selectors';
+import {selectSoilData} from 'terraso-mobile-client/store/selectors';
 import {updateSoilData} from 'terraso-mobile-client/model/soilId/soilIdSlice';
 
 import {FormInput} from 'terraso-mobile-client/components/form/FormInput';
