+++ dev-client/src/screens/SoilScreen/components/SoilSurfaceStatus.tsx
@@ -20,7 +20,7 @@ import {useTranslation} from 'react-i18next';
 import {
   selectSoilData,
   useSiteProjectSoilSettings,
-} from 'terraso-mobile-client/model/selectors';
+} from 'terraso-mobile-client/store/selectors';
 
 import {DataInputSummary} from 'terraso-mobile-client/components/DataInputSummary';
 import {
